+++ Specs/Scene/PrimitiveSpec.js
@@ -131,6 +131,7 @@ defineSuite([
         expect(primitive.allowPicking).toEqual(true);
         expect(primitive.asynchronous).toEqual(true);
         expect(primitive.debugShowBoundingVolume).toEqual(false);
+        expect(primitive.compressVertices).toEqual(true);
     });
 
     it('releases geometry instances when releaseGeometryInstances is true', function() {
@@ -783,7 +784,9 @@ defineSuite([
     it('renders when using asynchronous pipeline', function() {
         var primitive = new Primitive({
             geometryInstances : rectangleInstance1,
-            appearance : new PerInstanceColorAppearance()
+            appearance : new PerInstanceColorAppearance({
+                flat : true
+            })
         });
 
         frameState.camera.viewRectangle(rectangle1);